mafs-core-data/src/test/java/com/mikealbert/data/dao/CityZipCodeDAOTest.java
package com.mikealbert.data.dao;

import static org.junit.Assert.*;

import javax.annotation.Resource;

import org.junit.Test;
import org.springframework.data.domain.PageRequest;
import com.mikealbert.testing.BaseTest;
import com.mikealbert.data.dao.CityZipCodeDAO;
import com.mikealbert.data.entity.CityZipCodePK;


public class CityZipCodeDAOTest extends BaseTest {

//	@Resource
//	CityZipCodeDAO cityZipCodeDAO;
	
	@Test
	public void testGetOneCityGeoCodeForCanada(){
		final String GEOCODE = "0094";
		
		CityZipCodePK pk = new CityZipCodePK("CN", "BC", "003", "IOCO", "V3H");
				
//		assertEquals(cityZipCodeDAO.findById(pk).orElse(null).getGeoCode(), GEOCODE);
		assertEquals(0, 0);
		
	}
	
	@Test
	public void testGetCityGeoCodesForCanada(){
		final String COUNTRY_CODE = "CN";
		final int SIZE = 4;
		
		PageRequest page = new PageRequest(0,SIZE);
		
//		assertEquals(cityZipCodeDAO.findCityZipCodesByCountry(COUNTRY_CODE, page).size(), SIZE );
		assertEquals(0, 0);
	}

	@Test
	public void testGetListOfCitiesForZipCode(){
		final String ZIP = "45014";
		final int SIZE = 16;
		
//		assertEquals(cityZipCodeDAO.findCityZipCodesByZipCode(ZIP).size(), SIZE);
		assertEquals(0, 0);
	}

	
}
